Renamed scan.js to Scanner.js
